+++ back/models/wishPaymentItem.js
@@ -35,7 +35,7 @@ module.exports = class WishPaymentItem extends Model {
           allowNull: false,
         },
         qnt: {
-          type: DataTypes.INTEGER, // 수량
+          type: DataTypes.FLOAT, // 수량
           defaultValue: 1,
           allowNull: false, // 필수
         },
